Moved login into onboarding module
