+++ .idea/gradle.xml
@@ -18,10 +18,8 @@
             <option value="$PROJECT_DIR$/data" />
             <option value="$PROJECT_DIR$/data/onboarding" />
             <option value="$PROJECT_DIR$/domain" />
-            <option value="$PROJECT_DIR$/domain/authentication" />
             <option value="$PROJECT_DIR$/domain/onboarding" />
             <option value="$PROJECT_DIR$/ui" />
-            <option value="$PROJECT_DIR$/ui/authentication" />
             <option value="$PROJECT_DIR$/ui/onboarding" />
           </set>
         </option>
