+++ app/src/main/java/com/amalitech/bookmeetingroom/MainActivity.kt
@@ -10,8 +10,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.authentication.login.LoginScreen
 import com.example.core_ui.ui.theme.BookMeetingRoomTheme
+import com.example.onboarding_presentation.login.LoginScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
