--- domain/authentication/build.gradle.kts
@@ -1,7 +1,0 @@
-apply {
-    from("$rootDir/base-module.gradle")
-}
-
-dependencies {
-    "implementation"(project(Modules.core))
-}
