--- domain/authentication/src/main/java/com/example/authentication/use_case/LoginUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.authentication.use_case
-
-data class LoginUseCase(
-    val logIn: LogIn,
-    val validateEmail: ValidateEmail,
-    val validatePassword: ValidatePassword
-)
