+++ domain/onboarding/src/main/java/com/example/oboarding_domain/login/use_case/LoginUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.oboarding_domain.login.use_case
+
+data class LoginUseCase(
+    val logIn: LogIn,
+    val validateEmail: ValidateEmail,
+    val validatePassword: ValidatePassword
+)
