+++ settings.gradle
@@ -22,5 +22,3 @@ include ':core_ui'
 include ':domain:onboarding'
 include ':data:onboarding'
 include ':ui:onboarding'
-include ':ui:authentication'
-include ':domain:authentication'
