--- ui/authentication/build.gradle.kts
@@ -1,9 +1,0 @@
-apply {
-    from("$rootDir/compose-module.gradle")
-}
-
-dependencies {
-    "implementation"(project(Modules.core))
-    "implementation"(project(Modules.coreUI))
-    "implementation"(project(Modules.domainAuthentication))
-}
