--- ui/authentication/src/main/java/com/example/authentication/di/uiAuthenticationModule.kt
@@ -1,11 +1,0 @@
-package com.example.authentication.di
-
-import com.example.authentication.login.LoginViewModel
-import com.example.authentication.use_case.LoginUseCase
-import org.koin.dsl.module
-
-val uiAuthenticationModule = module {
-    single {
-        LoginViewModel(get())
-    }
-}
