+++ ui/onboarding/src/main/java/com/example/onboarding_presentation/login/LoginUiState.kt
@@ -1,0 +1,11 @@
+package com.example.onboarding_presentation.login
+
+import com.amalitech.core.util.UiText
+
+data class LoginUiState(
+    val email: String = "",
+    val password: String = "",
+    val error: UiText? = null,
+    val finishedLoggingIn: Boolean = false,
+    val snackBarValue: UiText? = null
+)
