+++ ui/onboarding/src/main/res/drawable/baseline_visibility_24.xml
