+++ ui/onboarding/src/main/res/drawable/baseline_visibility_off_24.xml
